Embed the configured chunk size in the LOB metadata, since it's not possible to validate a LOB in all cases without knowing this
This is because if the chunk size at the time the LOB was encoded is not known, you need the final chunk size to be known to derive the size of the other chunks, but if the last chunk file is missing or (worse) bad data, you can't calculate it correctly. For single-chunk files it's unnecessary but it's a very useful confirmation of the configuration on storage.