Fixing the case where client will endlessly try to push commits where it doesn't have data locally, but it's already on remote
For example, if you try to push just after cloning but not fetching, because at this point there is no record of commits having been pushed
It might seem like the solution is to mark commits as pushed after fetching, but this isn't reliable; when a commit is marked as pushed it
implies all its ancestors have been pushed too. Because fetch often only retrieves *partial* data (recent commits), it won't touch earlier
commits so can't vouch for them.

The degenerate case is if a client had unpushed binaries of its own, then didn't fetch for say a couple of months; the next fetch would only
download recent commits and leave a 'gap' - yet the unpushed LOBs would be in commits before that recent window. So if the fetch marked
data as pushed, the next push wouldn't look back far enough to pick up the unpushed binaries. It's perhaps unlikely but definitely could
happen on occasionally used repos.

Therefore instead we only ever mark LOBs on commits to be pushed during the push command. Where files are missing locally, we now check to
see if the missing ones are already on the remote. If so, we don't treat that commit as incomplete and will mark it as pushed (after any other SHAs
the remote doesn't have but the client has are pushed, if any). This means that there's a bit more work to do to confirm that the remote has these
resources, and it might be especially noticeable on the first push after cloning, since every commit will be checked. However this is the only
reliable way to avoid the scenario above, without the user manually saying that they definitely have nothing new to push.

An ideal time to improve this is when adding a clone wrapper. We know that directly after cloning there's not thing new to push, so could record all commits
as being pushed at that point. We could expose this as a lower-level command for manual use too.