Major refactor of push to use new push state walk methods
Also change how push works so that it estimates, pushes and marks pushed for each ref in turn rather than building up the list for all refs. We can do this now because push state respects ancestry and is always safe, and it results in a more ref-localised operation so you can have one ref completed and the other pending if necessary.